backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==44.0.0
httpx==0.28.1
websockets==14.1
python-multipart==0.0.19
pyotp==2.9.0
pydantic[email]

# ML / Data Science
numpy
pandas
scikit-learn
xgboost
lightgbm
optuna
hmmlearn
joblib
apscheduler
torch

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
